Update Carousel setup
[#185302492]
