+++ src/components/featuredPlaylists/styles.module.less
@@ -1,0 +1,17 @@
+@import '~@boclips-ui/styles/src/colors';
+
+.carousel__inner-slide {
+  width: calc(100% - 20px);
+  height: calc(100% - 20px);
+  left: 10px;
+  top: 10px;
+}
+
+.carouselWrapper {
+  background: @blue-100;
+  border-radius: 4px;
+  max-height: 370px;
+  padding: 1rem;
+  position: relative;
+
+}
