--- src/components/playlists/PromotedPlaylists.less
@@ -1,6 +1,0 @@
-.carousel__inner-slide {
-  width: calc(100% - 20px);
-  height: calc(100% - 20px);
-  left: 10px;
-  top: 10px;
-}
